+++ sql/core/src/test/scala/org/apache/spark/sql/DatasetUnpivotSuite.scala
@@ -17,7 +17,6 @@
 
 package org.apache.spark.sql
 
-import org.apache.spark.sql.errors.QueryErrorsSuiteBase
 import org.apache.spark.sql.functions.{length, struct, sum}
 import org.apache.spark.sql.test.SharedSparkSession
 import org.apache.spark.sql.types._
@@ -26,7 +25,6 @@ import org.apache.spark.sql.types._
  * Comprehensive tests for Dataset.unpivot.
  */
 class DatasetUnpivotSuite extends QueryTest
-  with QueryErrorsSuiteBase
   with SharedSparkSession {
   import testImplicits._
